requirements.txt
datasets
google-adk 